+++ lib/models/app_settings.g.dart
@@ -18,6 +18,8 @@ class AppSettingsAdapter extends TypeAdapter<AppSettings> {
     };
     return AppSettings(
       cardAnimations: fields[0] == null ? true : fields[0] as bool,
+      listeningInReviewName:
+          fields[2] == null ? 'sometimes' : fields[2] as String,
       updatedAt: fields[1] as DateTime?,
     );
   }
@@ -25,11 +27,13 @@ class AppSettingsAdapter extends TypeAdapter<AppSettings> {
   @override
   void write(BinaryWriter writer, AppSettings obj) {
     writer
-      ..writeByte(2)
+      ..writeByte(3)
       ..writeByte(0)
       ..write(obj.cardAnimations)
       ..writeByte(1)
-      ..write(obj.updatedAt);
+      ..write(obj.updatedAt)
+      ..writeByte(2)
+      ..write(obj.listeningInReviewName);
   }
 
   @override